% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds_dyn}
\alias{bounds_dyn}
\title{Calculate Tom Fleming monitoring guideline boundaries}
\usage{
bounds_dyn(events, power = 0.9, t1error = 0.025, lhr_null = NULL)
}
\arguments{
\item{events}{Vector. A vector with the planned number of events at the
analysis stages; `events` should be at least 1.}

\item{power}{Scalar. Power for the analysis. Default is 0.9.}

\item{t1error}{Scalar. One-sided Type-1 error. Default is 0.025.}

\item{lhr_null}{Scalar. The log hazard ratio (log-HR) threshold to rule out
under the null hypothesis at the final analysis satge (maximum information
level). Default is `NULL`. If `lhr_null` is `NULL`, then, the continuation
threshold at maximum information level is anchored at log-HR = 0 (that is a
hazard ratio equal to 1), and the log-HR threshold to rule out under the null
hypothesis at the final analysis is adjusted accordingly.}
}
\value{
List. A list object, that includes `lhr_con` (log-HR for
'continuation'), `lhr_null` (log-HR under the null hypothesis), `lhr_alt`
(log-HR under the alternative hypothesis) and `CI_level_monit_nullmax`
(levels of two-sided CIs which can be used to rule out lhr_null at each
analysis. Also, prints a table with the computed hazard ratio thresholds.
}
\description{
Calculate the overall survival hazard ratio thresholds at each
interim analysis stage according to the Tom Fleming monitoring guideline.
}
\examples{
res <- bounds_dyn(c(15, 30, 50, 69),
                   lhr_null = log(1.3))
}
