% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{flag_safety}
\alias{flag_safety}
\title{High level function to calculate the probability to ensure safety}
\usage{
flag_safety(method = "joint", logthres, events, loghr)
}
\arguments{
\item{method}{String. Either "joint" to calculate joint probabilities,
i.e., P(Reach analysis stage k and flag a safety issue) or "cond" to
calculate conditional probabilities, i.e., P(Flag a safety issue at stage k|
have not flagged a safety issue at any precedent analysis 1, ..., k-1).
Method "joint" is appropriate to explore at analysis stage, whereas "cond"
may be explored when the trial is actually in progress.}

\item{logthres}{Vector. The log hazard ratio thresholds for continuation
generated by the `bounds` function.}

\item{events}{Vector.  A vector with the planned number of events at the
analysis stages; `events` should be at least 1.}

\item{loghr}{Scalar. The assumed true log hazard ratio. Notice that a hazard
ratio larger than 1, would favor control (higher survival rates in
control) whereas a hazard ratio less than 1 would favor the treatment arm.}
}
\value{
Vector. Returns a vector with the computed probabilities at each
analysis stage.
}
\description{
This function calculates joint and conditional probabilities
of not flagging a safety issue at different analysis stages, assuming that
the Tom Fleming monitoring guideline is in place.
}
\examples{
logthres <- log(c(1.3, 1))
events <- c(50, 100)
loghr <- log(1)
res1 <- flag_safety(method='joint', logthres, events, loghr)
res2 <- flag_safety(method='cond', logthres, events, loghr)
}
