% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{flagprob}
\alias{flagprob}
\title{Probability to declare safety issue but not stop the trial}
\usage{
flagprob(logthres1, logthres2, events, loghr = log(1))
}
\arguments{
\item{logthres1}{Vector. The log hazard ratio thresholds for flagging a
safety issue.}

\item{logthres2}{Vector. The log hazard ratio thresholds for stopping the
trial due to safety issue.}

\item{events}{Vector.  A vector with the planned number of events at the
analysis stages; `events` should be at least 1.}

\item{loghr}{Scalar. The assumed true log hazard ratio. Notice that a hazard
ratio larger than 1, would favor control (higher survival rates in
control) whereas a hazard ratio less than 1 would favor the treatment arm.}
}
\value{
Vector. Returns a vector with the computed probabilities at each
analysis stage.
}
\description{
Given some pre-specified stopping boundaries, this function to
compute the joint probability to reach stage k and declare a safety issue
but not stop the trial given the assumed true hazard ratio.
}
\examples{
logthres1 <- log(c(1.3, 1))
logthres2 <- log(c(1.8, 1.3))
events <- c(50, 100)
loghr <- log(1)
res <- flagprob(logthres1, logthres2, events, loghr)
}
