% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrex.R
\name{lrexs}
\alias{lrexs}
\title{Standard exact test for time-to-event data}
\usage{
lrexs(data)
}
\arguments{
\item{data}{Tibble. Users need to input the time-to-event data, ensuring
that variables time, status and group include the required survival
information. Time-to-event data must be inserted with status, 0
if censored or 1 if dead; time and group, 0 for control and 1 for treatment
arm, variables.}
}
\value{
Scalar. The exact test p-value, i.e., Pr(LR < crit | HR) across all
permutations, where LR is the log-rank test statistic given the specific
permuted data set.
}
\description{
Calculate the exact test p-value based on the observed number of
events of a two-arm trial, assuming a true hazard ratio of 1 under the null
hypothesis.
}
\examples{
set.seed(12345)
n <- 20
dt <- tibble::tibble(id = 1:n,
                  time = rexp(n),
                  status = rbinom(n, size = 1, prob = 0.2),
                  group = rbinom(n, size = 1, prob = 0.5)
)
lrdt <- survival::survdiff(survival::Surv(time, status) ~ group, data=dt)

monitOS:::lrexs(dt)
}
