% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitOS-package.R
\docType{package}
\name{monitOS-package}
\alias{monitOS}
\alias{monitOS-package}
\title{monitOS: Monitoring OS in chronic or indolent cancers}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Evanthia Koukouli \email{evanthia.koukouli@novartis.com}

Authors:
\itemize{
  \item Thibaud Coroller \email{tibaud.coroller@novartis.com}
  \item Lisa Hampson \email{lisa.hampson@novartis.com}
}

}
\keyword{internal}
