% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprob}
\alias{postprob}
\title{Function to calculate the posterior probability of observing an inserted
number of events pattern for the treatment and control arms given prior
knowledge on each arm event rates  (Beta-Binomial model).}
\usage{
postprob(
  method = "beta",
  param0,
  param1,
  events0,
  events1,
  time0,
  time1,
  n0,
  n1,
  delta
)
}
\arguments{
\item{method}{String. Method for calculating the posterior probabilities. Two
methods are provided, 'beta' (Beta-Binomial conjugate family) and 'gamma'
(Gamma-Poisson conjugate family).}

\item{param0}{Vector. Shape parameters for the Beta prior distribution for
the control arm.}

\item{param1}{Vector. Shape parameters for the Beta prior distribution for
the treatment arm.}

\item{events0}{Scalar. Number of events in the control arm.}

\item{events1}{Scalar. Number of events in the treatment arm.}

\item{time0}{Scalar. Total subject's exposure time up to t in the control arm.
Default is NULL. Leave as NULL if method is Beta.}

\item{time1}{Scalar. Total subject's exposure time up to t in the treatment arm.
Default is NULL. Leave as NULL if method is Beta.}

\item{n0}{Scalar. Total number of patients in the control arm.}

\item{n1}{Scalar. Total number of patients in the treatment arm.}

\item{delta}{Scalar. Risk ratio difference under the null hypothesis.}
}
\value{
Scalar. Returns the posterior probability of observing the inserted
number of events pattern.
}
\description{
Function to calculate the posterior probability of observing an inserted
number of events pattern for the treatment and control arms given prior
knowledge on each arm event rates  (Beta-Binomial model).
}
\examples{
postrr <- postprob(param0 = c(1, 1), param1 = c(1, 1),
events0 = 2, events1 = 1, n0 = 100, n1 = 100, delta = 1, method = 'beta')
postrr

postrr <- postprob(param0 = c(1, 1), param1 = c(1, 1), time0 = 10, time1 = 1,
events0 = 2, events1 = 1, n0 = 100, n1 = 100, delta = 1, method = 'gamma')
postrr
}
