% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{safprob}
\alias{safprob}
\title{Joint probability to reach stage k and don't declare a safety issue}
\usage{
safprob(logthres, events, loghr)
}
\arguments{
\item{logthres}{Vector. The log hazard ratio thresholds for continuation
generated by the `bounds` function.}

\item{events}{Vector.  A vector with the planned number of events at the
analysis stages; `events` should be at least 1.}

\item{loghr}{Scalar. The assumed true log hazard ratio. Notice that a hazard
ratio larger than 1, would favor control (higher survival rates in
control) whereas a hazard ratio less than 1 would favor the treatment arm.}
}
\value{
Vector. Returns a vector with the computed joint probabilities at each
analysis stage.
}
\description{
Calculate the joint probability to reach stage k and don't
declare a safety issue given the assumed hazard ratio.
}
\examples{
logthres <- log(c(1.3, 1))
events <- c(50, 100)
loghr <- log(1)
res <- safprob(logthres, events, loghr)
}
