% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{stopprob}
\alias{stopprob}
\title{Function to the joint P{Reach stage k and declare a safety issue |
log-HR = theta}.}
\usage{
stopprob(logthres, events, loghr)
}
\arguments{
\item{logthres}{The log hazard ratio thresholds for continuation.}

\item{events}{A vector with the number of events at analysis stages.}

\item{loghr}{The true log hazard ratio.}
}
\value{
(vector): Returns a vector with the computed stopping probabilities
at each analysis stage.
}
\description{
Function to the joint P{Reach stage k and declare a safety issue |
log-HR = theta}.
}
\examples{
logthres <- log(c(1.3, 1))
events <- c(50, 100)
loghr <- log(1)
res <- stopprob(logthres, events, loghr)

}
