% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_bounds2.R
\name{tf_bounds2}
\alias{tf_bounds2}
\title{Compute the overall survival hazard ratio continuation threshold at each
interim analysis stage according to Version 1 of the Tom Fleming procedure.}
\usage{
tf_bounds2(events, power = 0.9, t1error = 0.025, low = FALSE, nullfin = NULL)
}
\arguments{
\item{events}{A vector with the planned number of events at the analysis
stages. It should be at least 1.}

\item{power}{Power for the analysis. A float ranging from 0 to 1. Default is
0.9.}

\item{t1error}{One sided type-1 error. A float ranging from 0 to 1. Default
is 0.025.}

\item{low}{A logical object showing whether the trial will reach a small OS
number of events at final analysis. We recommend to set as TRUE if the trial
yields less than 220 events. If TRUE, nullfin should be specified.}

\item{nullfin}{The log-HR threshold to rule out under the null hypothesis at
the final analysis in case of low number of events (<220). Default is NULL.}
}
\value{
A list object, that includes lhr_con, lhr_null and lhr_alt. Also,
prints a table with the computed thresholds
}
\description{
Compute the overall survival hazard ratio continuation threshold at each
interim analysis stage according to Version 1 of the Tom Fleming procedure.
}
\examples{
events1 <- c(68, 135, 220, 340) # sample size at each analysis stage
procs <- tf_bounds2(events1)

events2 <- c(15, 30, 50, 69)
procs <- tf_bounds2(events1, low = TRUE, nullfin = log(1.33))

}
