% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_ocs.R
\name{tf_ocs}
\alias{tf_ocs}
\title{Function to assess the operational characteristics of Tom Fleming's overall
survival monitoring guideline. The function gets as input the thresholds for
continuation at different stages of the study and returns the operational
characteristics of the monitoring rule under varying hazard ratio scenarios
inserted by the user.}
\usage{
tf_ocs(thres1, thres2 = NULL, events, hrr, hrs, col = NULL)
}
\arguments{
\item{thres1}{The hazard ratio thresholds to flag a safety issue at each
trial stage.}

\item{thres2}{The hazard ratio thresholds to flag a safety issue and stop the
trial. Default is NULL. Leave to default if specific stopping boundaries are
not known. Notice that thres2 is expected to be greater to thres1 for all
trial stages.}

\item{events}{A vector with the number of events at the specified analysis
stages.}

\item{hrr}{The range of hazard ratios to calculate the overall probability of
flagging a safety concern.}

\item{hrs}{The specific hazard ratio values to calculate stopping
probabilities.}

\item{col}{(Optional) A character vector with color names for the stopping
probabilities plot, must have same length to hrs.}
}
\value{
(list): Returns a list with two dataframes and three plots. The
dataframes include information on the stopping probabilities based on the
scenarios specified by the user and the plots include visualizations of these
results.
}
\description{
Function to assess the operational characteristics of Tom Fleming's overall
survival monitoring guideline. The function gets as input the thresholds for
continuation at different stages of the study and returns the operational
characteristics of the monitoring rule under varying hazard ratio scenarios
inserted by the user.
}
\examples{
thres1 <- c(1.3, 1)
events <- c(50, 100)
hrr <- seq(0.6, 1.5, by = 0.01)
hrs <- 1.3
res <- fleming_ocs(thres1, events, hrr, hrs)

thres1 <- c(1.3, 1)
thres2 <- c(1.8, 1.3)
events <- c(50, 100)
hrr <- seq(0.6, 1.5, by = 0.01)
hrs <- 1.3
res <- fleming_ocs(thres1, thres2, events, hrr, hrs)
}
