% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{Shiny app server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{generic shiny var}

\item{output}{generic shiny var}

\item{session}{generic shiny var}
}
\description{
Shiny app server
}
