% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{flag_safety}
\alias{flag_safety}
\title{High level function to calculate either joint or conditional safety
probabilities.}
\usage{
flag_safety(method = "joint", logthres, events, loghr)
}
\arguments{
\item{method}{String. Choice between joint or conditional.}

\item{logthres}{Vector. The log hazard ratio thresholds for continuation.}

\item{events}{Vector. A vector with the number of events at each analysis
stage.}

\item{loghr}{Scalar. The true log hazard ratio.}
}
\value{
Vector. Returns a vector with the computed probabilities at each
analysis stage.
}
\description{
High level function to calculate either joint or conditional safety
probabilities.
}
\examples{
logthres <- log(c(1.3, 1))
events <- c(50, 100)
loghr <- log(1)
res1 <- flag_safety(method='joint', logthres, events, loghr)
res2 <- flag_safety(method='cond', logthres, events, loghr)
}
