% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{flagprob}
\alias{flagprob}
\title{Function to compute the joint P{Reach stage k and declare a safety issue
but not stop the trial| log-HR = theta}.}
\usage{
flagprob(logthres1, logthres2, events, loghr = log(1))
}
\arguments{
\item{logthres1}{Vector. The log hazard ratio thresholds for flagging a
safety issue.}

\item{logthres2}{Vector. The log hazard ratio thresholds for stopping the
trial.}

\item{events}{Vector. A vector with the number of events at each analysis
stage.}

\item{loghr}{Scalar. The true log hazard ratio.}
}
\value{
Vector. Returns a vector with the computed probabilities at each
analysis stage.
}
\description{
Function to compute the joint P{Reach stage k and declare a safety issue
but not stop the trial| log-HR = theta}.
}
\examples{
logthres1 <- log(c(1.3, 1))
logthres2 <- log(c(1.8, 1.3))
events <- c(50, 100)
loghr <- log(1)
res <- flagprob(logthres1, logthres2, events, loghr)
}
