% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{flagprob}
\alias{flagprob}
\title{Function to the joint P{Reach stage k and declare a safety issue but not
stop the trial| log-HR = theta}.}
\usage{
flagprob(logthres1, logthres2, events, loghr)
}
\arguments{
\item{logthres1}{The log hazard ratio thresholds for flagging a safety issue.}

\item{logthres2}{The log hazard ratio thresholds for stopping the trial.}

\item{events}{A vector with the number of events at analysis stages.}

\item{loghr}{The true log hazard ratio.}
}
\value{
(vector): Returns a vector with the computed stopping probabilities
at each analysis stage.
}
\description{
Function to the joint P{Reach stage k and declare a safety issue but not
stop the trial| log-HR = theta}.
}
\examples{
logthres1 <- log(c(1.3, 1))
logthres2 <- log(c(1.8, 1.3))
events <- c(50, 100)
loghr <- log(1)
res <- stopprob(logthres1, logthres2, events, loghr)

}
