% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrex.R
\name{lrex}
\alias{lrex}
\title{Perform exact test given time-to-event data}
\usage{
lrex(method = "standard", data = NULL, n_perm = NULL, seed = NULL)
}
\arguments{
\item{method}{String. The `method` argument can be either 'standard' or
'heinze'. Default is 'standard'.}

\item{data}{Tibble. Users need to input the time-to-event data, ensuring
that variables time, status and group include the required survival
information. Time-to-event data must be inserted with status, 0
if censored or 1 if dead; time and group, 0 for control and 1 for treatment
arm, variables.}

\item{n_perm}{Scalar. The number of permutations to be performed. Default is
`NULL`. Do not change, if `method = "standard"`.}

\item{seed}{Scalar. The seed for random number generation. Default is
`NULL`. Do not change, if `method = "standard"`.}
}
\value{
Scalar. The exact test p-value, i.e., Pr(LR < crit | HR) across all
permutations, where LR is the log-rank test statistic given the specific
permuted data set.
}
\description{
Perform the exact test based on time-to-event data for a two-arm
trial. We are offering two methods, the standard exact test and the exact
test for unequal follow-up time, as alternatives to the log-rank test when
the number of events is very small, e.g., less than 25.
}
\examples{
set.seed(12345)
n <- 50
dt <- tibble::tibble(id = 1:n,
                     time = rexp(n),
                     status = c(rbinom(n/2, size = 1, prob = 0.2),
                                rbinom(n/2, size = 1, prob = 0.1)),
                     group = rep(0:1, each = n/2)
)
table(ifelse(dt$status==1, "dead", "censored") , dt$group)

# standard log-rank test
lrdt <- survival::survdiff(survival::Surv(time, status) ~ group, data=dt)
lrdt$pvalue

n_perm <- 1000
data <- dt
seed <- 12345
# method = "standard" is performed by default
lrex(data = data)
# method = "heinze" to consider follow-up times
lrex(method = "heinze", data = data, n_perm = n_perm, seed = seed)
}
