% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrex.R
\name{lrex}
\alias{lrex}
\title{Function to perform the exact test for a two-arm trial.}
\usage{
lrex(
  method = "standard",
  events = NULL,
  n0 = NULL,
  n1 = NULL,
  crit = NULL,
  data = NULL,
  n_perm = NULL,
  seed = 1234
)
}
\arguments{
\item{method}{String. Method can be either "standard" or "heinze". Default is
"standard", calculating the exact test p-value given the data provided. The
standard exact test inputs the number of observed events along with
the total sample size per arm and the log-rank test statistic value. If
"heinze" method is selected, users need to input the time-to-event data and
the number of permutations to be performed.}

\item{events}{Scalar. The total number of observed events.}

\item{n0}{Scalar. The sample size in the control group.}

\item{n1}{Scalar. Sample size in the treatment group.}

\item{crit}{Scalar. The observed critical value of the log-rank test.}

\item{data}{Tibble. Time-to-event data must be inserted with status, 0
if censored or 1 if dead; time and group variables.}

\item{n_perm}{Scalar. The number of permutations to be performed.}

\item{seed}{Scalar. The seed for the permutations.}
}
\value{
Scalar. The exact test p-value, i.e., Pr(LR < crit | HR) across all
permutations where LR is the log-rank test result given the specific
permuted data set.
}
\description{
Function to perform the exact test for a two-arm trial.
}
\examples{
set.seed(12345)
n <- 20
dt <- tibble::tibble(id = 1:n,
                     time = rexp(n),
                     status = rbinom(n, size = 1, prob = 0.2),
                     group = rbinom(n, size = 1, prob = 0.5)
)
# 3 events occurred
# compute log-rank test result
lrdt <- survival::survdiff(survival::Surv(time, status) ~ group, data=dt)

method <- "standard"
events <- sum(dt$status)
n0 <- sum(dt$group==0)
n1 <- sum(dt$group==1)
crit <- lrdt$chisq
lrex(method = method,
     events = events,
     n0 = n0,
     n1 = n1,
     crit = crit)

n_perm <- 100
lrex(method = "heinze", data = dt, n_perm = n_perm, seed = 12345)

}
