% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrex.R
\name{lrexs}
\alias{lrexs}
\title{Calculate the exact test p-value assuming a true HR=1 scenario and the
observed number of events in a two-arm trial.}
\usage{
lrexs(data, n_perm)
}
\arguments{
\item{data}{Tibble. Time-to-event data must be inserted with status, 0
if censored or 1 if dead; time and group variables.}

\item{n_perm}{Scalar. The number of permutations to be performed.}

\item{seed}{Scalar. The seed for the permutations.}
}
\value{
Scalar. The exact test p-value, i.e., Pr(LR < crit | HR) across all
permutations where LR is the log-rank test result given the specific
permuted data set.
}
\description{
Calculate the exact test p-value assuming a true HR=1 scenario and the
observed number of events in a two-arm trial.
}
\examples{
set.seed(12345)
n <- 20
dt <- tibble::tibble(id = 1:n,
                  time = rexp(n),
                  status = rbinom(n, size = 1, prob = 0.2),
                  group = rbinom(n, size = 1, prob = 0.5)
)
lrdt <- survival::survdiff(survival::Surv(time, status) ~ group, data=dt)

monitOS:::lrexs(dt, n_perm = 1e04)

}
