% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrex.R
\name{lrexu}
\alias{lrexu}
\title{Given unequal patient follow-up time, it gets the survival data and the
number of permutations to be performed and returns the probability that we
observe a p-value as extreme as the one we have observed based on the
log-rank test of the data.}
\usage{
lrexu(data, n_perm, seed)
}
\arguments{
\item{data}{Tibble. Time-to-event data must be inserted with status, 0
if censored or 1 if dead; time and group variables.}

\item{n_perm}{Scalar. The number of permutations to be performed.}

\item{seed}{Scalar. The seed for the permutations.}
}
\value{
Scalar. The exact test p-value, i.e., Pr(LR < crit | HR) across all
permutations where LR is the log-rank test result given the specific
permuted data set.
}
\description{
Given unequal patient follow-up time, it gets the survival data and the
number of permutations to be performed and returns the probability that we
observe a p-value as extreme as the one we have observed based on the
log-rank test of the data.
}
\examples{
set.seed(12345)
n <- 20 # number of patients
dt <- tibble::tibble(time = rexp(n),
                 status = rbinom(n, size = 1, prob = 0.2),
                 group = rbinom(n, size = 1, prob = 0.5)
)
monitOS:::lrexu(data = dt, n_perm = 100, seed = 12345)

}
