% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprob.R
\name{postprob}
\alias{postprob}
\title{Calculate posterior probabilities for safety monitoring stopping rule}
\usage{
postprob(
  method = "beta",
  param0,
  param1,
  events0,
  events1,
  time0,
  time1,
  n0,
  n1,
  delta
)
}
\arguments{
\item{method}{String. Method for calculating the posterior probabilities. Two
methods are provided, 'beta' (Beta-Binomial conjugate family) and 'gamma'
(Gamma-Poisson conjugate family). Default is set to 'beta'.}

\item{param0, param1}{Vectors with two values each. Parameter values for the
prior distributions of the control and treatment arms respectively.}

\item{events0, events1}{Scalars. Assumed number of events within the control
and treatment arms respectively.}

\item{time0, time1}{Scalars. Specify only if `method = 'gamma'`, otherwise
leave as `NULL`. This is the total subjects' exposure time in the control and
treatment arms. Default is `NULL`.}

\item{n0, n1}{Scalars. Total number of patients in the control and treatment
arms respectively.}

\item{delta}{Scalar. Assumed risk ratio difference under the null hypothesis.}
}
\value{
Scalar. Returns the posterior probability of observing the specified
number of events pattern.
}
\description{
Function to calculate the posterior probability of observing the
specified number of events pattern within a two-arm study given prior
knowledge on each arm's event rate distribution.
}
\examples{
postrr <- postprob(param0 = c(1, 1), param1 = c(1, 1),
events0 = 2, events1 = 1, n0 = 100, n1 = 100, delta = 1, method = 'beta')
postrr

postrr <- postprob(param0 = c(1, 1), param1 = c(1, 1), time0 = 10, time1 = 1,
events0 = 2, events1 = 1, n0 = 100, n1 = 100, delta = 1, method = 'gamma')
postrr
}
