% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{run_app}
\alias{run_app}
\title{monitOS app}
\usage{
run_app()
}
\value{
No return value, runs shiny app
}
\description{
Runs the shiny app to guide user choice adequate settings to calculate
the positivity thresholds to monitor overall survival (OS)
}
