% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_safety.R
\name{safprobc}
\alias{safprobc}
\title{Function to calculate the P{Reach stage k and not declare a safety issue |
log-HR = theta, didn't declare a safety issue at any previous stage}.}
\usage{
safprobc(logthres, events, loghr)
}
\arguments{
\item{logthres}{Vector. The log hazard ratio thresholds for continuation.}

\item{events}{Vector. A vector with the number of events at each analysis
stage.}

\item{loghr}{Scalar. The true log hazard ratio.}
}
\value{
Vector. Returns a vector with the computed conditional probabilities
at each analysis stage.
}
\description{
Function to calculate the P{Reach stage k and not declare a safety issue |
log-HR = theta, didn't declare a safety issue at any previous stage}.
}
\examples{
logthres <- log(c(1.3, 1))
events <- c(50, 100)
loghr <- log(1)
res <- safprobc(logthres, events, loghr)

}
